fixed height for suggestion
